@@ -71,14 +74,14 @@
     "@azure-tools/test-utils-vitest": "^1.0.0",
     "@azure/dev-tool": "^1.0.0",
     "@azure/eslint-plugin-azure-sdk": "^3.0.0",
-    "@azure/identity": "^4.6.0",
+    "@azure/identity": "^4.8.0",
     "@types/node": "^18.0.0",
     "@vitest/browser": "^3.0.9",
     "@vitest/coverage-istanbul": "^3.0.9",
     "dotenv": "^16.0.0",
     "eslint": "^9.9.0",
-    "playwright": "^1.50.0",
-    "typescript": "~5.6.2",
+    "playwright": "^1.51.1",
+    "typescript": "~5.7.2",
     "vitest": "^3.0.9"
   },
   "scripts": {
