@@ -1,58 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { ComputeScheduleContext as Client, OperationsListOptionalParams } from "../index.js";
-import {
-  _OperationListResult,
-  _operationListResultDeserializer,
-  Operation,
-} from "../../models/models.js";
-import {
-  PagedAsyncIterableIterator,
-  buildPagedAsyncIterator,
-} from "../../static-helpers/pagingHelpers.js";
-import {
-  StreamableMethod,
-  PathUncheckedResponse,
-  createRestError,
-  operationOptionsToRequestParameters,
-} from "@azure-rest/core-client";
-
-export function _listSend(
-  context: Client,
-  options: OperationsListOptionalParams = { requestOptions: {} },
-): StreamableMethod {
-  return context.path("/providers/Microsoft.ComputeSchedule/operations").get({
-    ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
-    queryParameters: { "api-version": context.apiVersion },
-  });
-}
-
-export async function _listDeserialize(
-  result: PathUncheckedResponse,
-): Promise<_OperationListResult> {
-  const expectedStatuses = ["200"];
-  if (!expectedStatuses.includes(result.status)) {
-    throw createRestError(result);
-  }
-
-  return _operationListResultDeserializer(result.body);
-}
-
-/** List the operations for the provider */
-export function list(
-  context: Client,
-  options: OperationsListOptionalParams = { requestOptions: {} },
-): PagedAsyncIterableIterator<Operation> {
-  return buildPagedAsyncIterator(
-    context,
-    () => _listSend(context, options),
-    _listDeserialize,
-    ["200"],
-    { itemName: "value", nextLinkName: "nextLink" },
-  );
-}
+export { list } from "./operations.js";
+export { OperationsListOptionalParams } from "./options.js";
